Switch to c++_static and make lots of fixes

+++ app/src/main/jni/desmume/src/android/7zip.cpp
@@ -350,7 +350,7 @@ int ArchiveFile::ExtractItem(int index, unsigned char* outBuffer, int bufSize) c
 			if(SUCCEEDED(object->Open(ifs,0,0)))
 			{
 				OutStream* os = new OutStream(index, outBuffer, item.size);
-				const UInt32 indices [1] = {index};
+				const UInt32 indices [1] = {static_cast<UInt32>(index)};
 				hr = object->Extract(indices, 1, 0, os);
 				object->Close();
 			}
@@ -399,7 +399,7 @@ int ArchiveFile::ExtractItem(int index, const char* outFilename) const
 			{
 				gameInfo.resize(rv);
 				OutStream* os = new OutStream(index, outFilename);
-				const UInt32 indices [1] = {index};
+				const UInt32 indices [1] = {static_cast<UInt32>(index)};
 				hr = object->Extract(indices, 1, 0, os);
 				object->Close();
 			}
